requirements.txt
trio
psycopg2-binary
python-dotenv
databases[asyncpg]
databases[aiosqlite]
fastapi<=0.111.0
uvicorn
sqlalchemy
greenlet
aiosqlite
itsdangerous
pytest
pydantic>=2.7.0
pydantic-settings
httpx
jinja2
apscheduler>=3.10.0

